Use LeadPreview from leftpane for Leads and LeadAdd
