+++ src/views/LeadAdd/LeadDetail/LeadForm/ApplyAll/index.js
