+++ src/views/LeadAdd/LeadDetail/LeadForm/ApplyAll/styles.scss
