+++ src/views/LeadAdd/LeadDetail/LeadPreviewWithTabular/LeadTabular/ExcelSettings/styles.scss
