+++ src/views/LeadAdd/LeadDetail/LeadPreviewWithTabular/LeadTabular/styles.scss
