--- src/views/LeadAdd/LeadPreview/styles.scss
@@ -1,48 +1,0 @@
-.gallery-file {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-
-    :global {
-        .urlbar {
-            flex-shrink: 0;
-        }
-
-        .doc-container {
-            display: flex;
-            flex-direction: column;
-            flex-grow: 1;
-
-            .gallery-docs {
-                display: flex;
-                flex-direction: column;
-                flex-grow: 1;
-            }
-
-            iframe {
-                flex-grow: 1;
-            }
-        }
-    }
-}
-
-.tabular-modal {
-    transition: .2s all ease-in-out;
-    width: auto;
-    overflow: auto;
-
-    :global {
-        .page-1 {
-            width: 50vw;
-            min-width: 320px;
-            height: auto;
-            min-height: 200px;
-            max-height: 60vh;
-        }
-    }
-
-    .tabular-book {
-        width: 96vw;
-        height: 92vh;
-    }
-}
