+++ src/views/Leads/LeadPreviewModal/styles.scss
